requirements.txt
discord.py>=2.3.2
yt_dlp>=2025.02.19
PyNaCl>=1.5.0
mutagen